refactor(trexconverter): move TypedString; add module-info
- Avoid JPMS split package with msv-core (com.sun.msv.grammar.trex)
- Add net.java.dev.msv.trexconverter module descriptor
- Drop obsolete TypedString import from TREXWriter
