+++ trexconverter/src/main/java/com/sun/msv/writer/trex/TREXWriter.java
@@ -11,7 +11,6 @@ package com.sun.msv.writer.trex;
 
 import com.sun.msv.grammar.*;
 import com.sun.msv.grammar.util.ExpressionWalker;
-import com.sun.msv.grammar.trex.TypedString;
 import com.sun.msv.grammar.relax.*;
 import com.sun.msv.datatype.SerializationContext;
 import com.sun.msv.datatype.xsd.*;
